+++ Dyalect/Compiler/CompilerError.cs
@@ -34,7 +34,7 @@
 
         StaticOnlyMethods = 215,
 
-        CodeIslandMultipleExpressions = 216,
+        // = 216,
 
         CodeIslandEmpty = 217,